куча исправлений и в гредле и в слое Data все связано с совместимостью

+++ app/src/main/java/com/example/menuwithcleanachitecture/Domain/models/DishItem.kt
@@ -6,7 +6,7 @@ data class DishItem(
     val url_dish: String,
     val name_dish: String,
     val price_dish: String,
-    val countPrice: Int,
+    val countPrice: Double,
     val weight_dish: String,
     val quantity: Int
